Cleaned the code, separated the models and methods. Added getTopRated
