+++ my-app/src/App.tsx
@@ -7,10 +7,11 @@ import Background from "./components/background";
 import { theme } from "./theme/theme";
 import callApi from "./middleware/callApi";
 import { ChangeEvent, useCallback, useEffect, useState } from "react";
-import { Movie } from "./models/movies";
+import { Movie } from "./models/Movie";
 import HomeContent from "./components/homeContent";
 import { searchMovies } from "./middleware/searchMovies";
 import { useDebounce } from "./middleware/useDebounce";
+// import {getTopRated} from "./middleware/getTopRated";
 // import {getCredits, getPerson, getPersonByID, searchPerson} from "./middleware/movieCredits";
 
 
@@ -44,12 +45,13 @@ function App() {
     // console.log(await getPersonByID(250))
     // console.log(await searchPerson("Bradley Cooper"))
     // console.log(await getPerson(250))
+    // console.log(await getTopRated())
   };
 
     useEffect(() => {
         initialLoad()
     },[])
-    console.log(movies)
+    // console.log(movies)
 
   return (
     <ThemeProvider theme={theme}>
