+++ my-app/src/middleware/getCredits.ts
@@ -1,0 +1,7 @@
+import {Credits} from "../models/Credits";
+import callApi from "./callApi";
+
+export const getCredits = async (movie_id:number):Promise<Credits> => {
+    return await callApi(`
+https://api.themoviedb.org/3/movie/${movie_id}/credits?api_key=ac1ccaf7cc1c015abd2c2cddca72cb16&language=en-US`,"GET")
+}
