+++ my-app/src/models/Cast.ts
@@ -1,0 +1,14 @@
+export interface Cast{
+    adult: boolean
+    gender:  number | null
+    id: number,
+    known_for_department: string,
+    name: string,
+    original_name: string
+    popularity:  number
+    profile_path:  string | null
+    cast_id: number,
+    character: string,
+    credit_id: string,
+    order: number
+}
