+++ my-app/src/models/Credits.ts
@@ -1,0 +1,8 @@
+import {Cast} from "./Cast";
+import {Crew} from "./Crew";
+
+export interface Credits{
+    id: number,
+    cast:Cast[],
+    crew: Crew[]
+}
