+++ my-app/src/models/Crew.ts
@@ -1,0 +1,13 @@
+export interface Crew{
+    adult:boolean
+    gender: number | null
+    id:number,
+    known_for_department: string
+    name: string
+    original_name: string
+    popularity: number
+    profile_path: string | null
+    credit_id: string
+    department: string
+    job: string
+}
