+++ my-app/src/models/MovieBox.ts
@@ -1,0 +1,6 @@
+export interface MovieBox {
+    id: number,
+    poster_path: string,
+    title: string,
+    vote_average: number,
+}
