+++ my-app/src/models/Movies.ts
@@ -1,0 +1,5 @@
+import {Movie} from "./Movie";
+
+export interface Movies {
+    movies:Movie[]
+}
