+++ my-app/src/models/Person.ts
@@ -1,0 +1,19 @@
+import {Movie} from "./Movie";
+
+export interface Person{
+    birthday: string | null,
+    known_for_department: string,
+    deathday: null | string,
+    id:  number,
+    name: string,
+    also_known_as: String[],
+    gender: number,
+    biography: string,
+    popularity: number,
+    place_of_birth: string | null,
+    profile_path:  string | null,
+    adult:  boolean,
+    imdb_id: string,
+    homepage: null | string,
+    known_for: Movie[]|{profile_path: string, adult: boolean, id: number, known_for: {}[], name: string, popularity: number}
+}
