+++ my-app/src/models/PersonByID.ts
@@ -1,0 +1,16 @@
+export interface PersonByID{
+    birthday: string | null,
+    known_for_department: string,
+    deathday: null | string,
+    id:  number,
+    name: string,
+    also_known_as: String[],
+    gender: number,
+    biography: string,
+    popularity: number,
+    place_of_birth: string | null,
+    profile_path:  string | null,
+    adult:  boolean,
+    imdb_id: string
+    homepage: null | string
+}
