+++ my-app/src/models/SMovie.ts
@@ -1,0 +1,39 @@
+export interface sMovie {
+  adult?: boolean,
+  backdrop_path?: string,
+  genre_ids?: number[],
+  id?: number,
+  original_language?: string,
+  original_title?: string,
+  overview?: string,
+  popularity?: number,
+  poster_path?: string,
+  release_date?: string,
+  title?: string,
+  video?: boolean,
+  vote_average?: number,
+  vote_count?: number,
+  media_type?:  string,
+  first_air_date?: string,
+  origin_country?:string[],
+  name?:string
+  original_name?:string
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
