+++ my-app/src/models/SearchMovie.ts
@@ -1,0 +1,8 @@
+import {Movie} from "./Movie";
+
+export interface SearchMovie{
+    page: number
+    results: Movie[]
+    total_pages:number
+    total_results:number
+}
