+++ my-app/src/models/SearchPerson.ts
@@ -1,0 +1,8 @@
+import {Result} from "./Result";
+
+export interface SearchPerson{
+    page:number,
+    results: Result[],
+    total_results:number,
+    total_pages:number,
+}
