+++ my-app/src/models/User.ts
@@ -1,0 +1,4 @@
+export interface User{
+    username:string,
+    password:string
+}
